Add files via upload
How to install Fortran compiler and install it on Windows 11 machine. Along with a simple Hello Fortran example which was written in Nano editor. 

The files are located in the installation folder of Msys. In my case I have installed it in my Program Files on File Explorer and also created a separate Fortran folder. So the path to all my Fortran files are as follows

C:\Program Files\msys64\home\YourWindowsUsername\fortran_folder